(feat) Implement handleApplicationCommand helper
